Switch to bundled tbb version "4.1 Update 2013/04/01"
  - Enable compilation with clang version 3.0 by making a cast in ASM code
    explicit

  - Suppress warnings about pointer-to-object to pointer-to-function casts

git-svn-id: https://svn.dealii.org/trunk@29365 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/bundled/tbb41_20130401oss/src/tbb/tools_api/ittnotify_config.h
@@ -264,7 +264,7 @@ INLINE int __TBB_machine_fetchadd4(volatile void* ptr, long addend)
     int result;
     __asm__ __volatile__("lock\nxaddl %0,%1"
                           : "=r"(result),"=m"(*(long*)ptr)
-                          : "0"(addend), "m"(*(long*)ptr)
+                          : "0"((int)addend), "m"(*(long*)ptr)
                           : "memory");
     return result;
 }
